refactor(workspace): trocear UnifiedSidebar en sidebar/ — fase 2 (03/T02)
UnifiedSidebar baja de 2.123 a 933 líneas extrayendo sin cambios de
comportamiento a app/components/workspace/sidebar/: sidebarHelpers.ts
(pickFolderColor/parseMeta/getFolderColor/buildTree/isDescendant + tipos),
SidebarContextMenu.tsx, SidebarModals.tsx (move/delete/new-folder/url),
SidebarFileTree.tsx (TreeNode + FileTree recursivos, juntos) y
AddResourceMenu.tsx. El componente principal conserva su API; imports
muertos podados.
