+++ app/components/workspace/UnifiedSidebar.tsx
@@ -1,40 +1,6 @@
 import { useState, useEffect, useCallback, useRef, useMemo, lazy, Suspense, type ReactNode } from 'react';
 const CloudFilePicker = lazy(() => import('@/components/cloud/CloudFilePicker'));
-import {
-  ChevronDown,
-  Search,
-  Settings,
-  Moon,
-  Sun,
-  Home,
-  Calendar,
-  BookOpen,
-  Tag,
-  Store,
-  Folder,
-  FileText,
-  RefreshCw,
-  X,
-  MoreHorizontal,
-  Edit3,
-  Trash2,
-  FolderInput,
-  FolderPlus,
-  Check,
-  Hash,
-  Plus,
-  NotebookPen,
-  Link,
-  Upload,
-  Cloud,
-  Bot,
-  Workflow,
-  Zap,
-  Activity,
-  Layers,
-  PanelRightOpen,
-  Maximize2,
-} from 'lucide-react';
+import { ChevronDown, Settings, Moon, Sun, Home, Calendar, BookOpen, Tag, Store, RefreshCw, FolderPlus, Plus, Bot, Workflow, Zap, Activity, Layers } from 'lucide-react';
 import { useTranslation } from 'react-i18next';
 import { useShallow } from 'zustand/react/shallow';
 import { useAppStore } from '@/lib/store/useAppStore';
@@ -57,1175 +23,19 @@ import {
   HUB_WORKFLOWS_CHANGED,
 } from '@/lib/hub/hubEvents';
 import { db } from '@/lib/db/client';
-import DomeResourceIcon from '@/components/ui/DomeResourceIcon';
 import ManyIcon from '@/components/many/ManyIcon';
-import { FOLDER_COLOR_OPTIONS, NAMED_FOLDER_COLORS } from '@/lib/ui/palettes';
+
 
 // ---------------------------------------------------------------------------
 // Folder colors — central palette in app/lib/ui/palettes.ts (persisted in DB)
 // ---------------------------------------------------------------------------
-const FOLDER_AUTO_PALETTE = FOLDER_COLOR_OPTIONS.map((o) => o.value);
-function pickFolderColor(): string {
-  return FOLDER_AUTO_PALETTE[Math.floor(Math.random() * FOLDER_AUTO_PALETTE.length)];
-}
 
-function parseMeta(resource: Resource): Record<string, unknown> {
-  const m = resource.metadata;
-  if (!m) return {};
-  if (typeof m === 'string') { try { return JSON.parse(m) as Record<string, unknown>; } catch { return {}; } }
-  return m as Record<string, unknown>;
-}
+// Subcomponentes extraídos (03/T02) — misma UI, archivos en ./sidebar/.
+import { pickFolderColor, parseMeta } from './sidebar/sidebarHelpers';
+import FileTree from './sidebar/SidebarFileTree';
+import { NewFolderModal, UrlInputModal } from './sidebar/SidebarModals';
+import AddResourceMenu from './sidebar/AddResourceMenu';
 
-function getFolderColor(resource: Resource): string {
-  const color = parseMeta(resource).color as string | undefined;
-  if (!color) return 'var(--dome-accent)';
-  if (color.startsWith('#')) return color;
-  return NAMED_FOLDER_COLORS[color] ?? 'var(--dome-accent)';
-}
-
-// ---------------------------------------------------------------------------
-// Tree data
-// ---------------------------------------------------------------------------
-interface TreeNodeData {
-  id: string;
-  name: string;
-  type: 'folder' | 'notebook' | 'url' | 'youtube' | 'pdf' | 'image' | 'audio' | 'video' | 'ppt' | 'file' | 'artifact';
-  children?: TreeNodeData[];
-  resource?: Resource;
-}
-
-function buildTree(resources: Resource[]): TreeNodeData[] {
-  const folderMap = new Map<string, TreeNodeData>();
-  const roots: TreeNodeData[] = [];
-  const folders = resources.filter((r) => r.type === 'folder');
-  const nonFolders = resources.filter((r) => r.type !== 'folder');
-
-  for (const f of folders) {
-    folderMap.set(f.id, { id: f.id, name: f.title, type: 'folder', children: [], resource: f });
-  }
-  for (const r of nonFolders) {
-    const node: TreeNodeData = { id: r.id, name: r.title, type: r.type as TreeNodeData['type'], resource: r };
-    if (r.folder_id && folderMap.has(r.folder_id)) folderMap.get(r.folder_id)!.children!.push(node);
-    else roots.push(node);
-  }
-  for (const f of folders) {
-    const node = folderMap.get(f.id)!;
-    if (f.folder_id && folderMap.has(f.folder_id)) folderMap.get(f.folder_id)!.children!.push(node);
-    else roots.push(node);
-  }
-  return roots;
-}
-
-// ---------------------------------------------------------------------------
-// Context menu
-// ---------------------------------------------------------------------------
-interface CtxState {
-  visible: boolean;
-  x: number;
-  y: number;
-  resource: Resource | null;
-}
-
-interface ContextMenuProps {
-  state: CtxState;
-  onClose: () => void;
-  onRename: (r: Resource) => void;
-  onMove: (r: Resource) => void;
-  onColorChange: (r: Resource, color: string) => void;
-  onDelete: (r: Resource) => void;
-  onNewFolder: (parentId: string | null) => void;
-  onOpenInSplit?: (r: Resource) => void;
-  onOpenInWindow?: (r: Resource) => void;
-  /**
-   * True when the user has an active tab that can host a split view
-   * (any non-home tab where `openResourceInSplit` is meaningful).
-   */
-  canOpenInSplit?: boolean;
-}
-
-function ContextMenu({
-  state,
-  onClose,
-  onRename,
-  onMove,
-  onColorChange,
-  onDelete,
-  onNewFolder,
-  onOpenInSplit,
-  onOpenInWindow,
-  canOpenInSplit,
-}: ContextMenuProps) {
-  const { t } = useTranslation();
-  const menuRef = useRef<HTMLDivElement>(null);
-  const [showColors, setShowColors] = useState(false);
-  const [hoveredColor, setHoveredColor] = useState<string | null>(null);
-
-  useEffect(() => {
-    if (!state.visible) return;
-    setShowColors(false);
-    setHoveredColor(null);
-    const handle = (e: MouseEvent) => { if (menuRef.current && !menuRef.current.contains(e.target as Node)) onClose(); };
-    const handleKey = (e: KeyboardEvent) => { if (e.key === 'Escape') onClose(); };
-    document.addEventListener('mousedown', handle);
-    document.addEventListener('keydown', handleKey);
-    return () => { document.removeEventListener('mousedown', handle); document.removeEventListener('keydown', handleKey); };
-  }, [state.visible, onClose]);
-
-  if (!state.visible || !state.resource) return null;
-  const r = state.resource;
-  const isFolder = r.type === 'folder';
-  const currentColor = parseMeta(r).color as string | undefined;
-
-  const hoveredLabel = hoveredColor
-    ? (FOLDER_COLOR_OPTIONS.find((o) => o.value === hoveredColor)?.label ?? null)
-    : null;
-
-  const menuWidth = 196;
-  const left = Math.min(state.x, window.innerWidth - menuWidth - 8);
-  const estimatedHeight = isFolder ? 320 : 200;
-  const top = Math.min(state.y, window.innerHeight - estimatedHeight - 8);
-
-  return (
-    <div
-      ref={menuRef}
-      className="fixed z-[var(--z-popover)] overflow-hidden"
-      style={{
-        left,
-        top,
-        width: menuWidth,
-        background: 'var(--dome-surface)',
-        border: '1px solid var(--dome-border)',
-        borderRadius: 10,
-        boxShadow: '0 8px 24px rgba(0,0,0,0.14), 0 2px 6px rgba(0,0,0,0.08)',
-      }}
-    >
-      {/* Resource label */}
-      <div className="px-3 pt-2.5 pb-1.5" style={{ borderBottom: '1px solid var(--dome-border)' }}>
-        <p className="truncate" style={{ fontSize: 12, fontWeight: 600, letterSpacing: '0.04em', textTransform: 'uppercase', color: 'var(--dome-text-muted)' }}>
-          {isFolder ? 'Carpeta' : r.type === 'notebook' ? 'Cuaderno' : r.type === 'url' ? 'URL' : r.type === 'pdf' ? 'PDF' : 'Archivo'}
-        </p>
-        <p className="truncate mt-0.5" style={{ fontSize: 12.5, fontWeight: 500, color: 'var(--dome-text)' }}>{r.title}</p>
-      </div>
-
-      <div style={{ padding: '4px 0' }}>
-        {/* Open as reference / popout — only for non-folder resources */}
-        {!isFolder && onOpenInSplit && canOpenInSplit && (
-          <CtxItem
-            icon={<PanelRightOpen className="size-3.5" />}
-            label={t('focused_editor.open_reference', 'Abrir como referencia')}
-            onClick={() => { onOpenInSplit(r); onClose(); }}
-          />
-        )}
-        {!isFolder && onOpenInWindow && r.type === 'note' && (
-          <CtxItem
-            icon={<Maximize2 className="size-3.5" />}
-            label={t('focused_editor.popout', 'Abrir en ventana')}
-            onClick={() => { onOpenInWindow(r); onClose(); }}
-          />
-        )}
-        {!isFolder && (onOpenInSplit || onOpenInWindow) && (
-          <div style={{ height: 1, background: 'var(--dome-border)', margin: '4px 6px' }} />
-        )}
-
-        {/* Rename */}
-        <CtxItem icon={<Edit3 className="size-3.5" />} label="Renombrar" onClick={() => { onRename(r); onClose(); }} />
-
-        {/* Move */}
-        <CtxItem icon={<FolderInput className="size-3.5" />} label="Mover a carpeta" onClick={() => { onMove(r); onClose(); }} />
-
-        {/* New subfolder — folders only */}
-        {isFolder && (
-          <CtxItem icon={<FolderPlus className="size-3.5" />} label="Nueva subcarpeta" onClick={() => { onNewFolder(r.id); onClose(); }} />
-        )}
-
-        {/* Color picker — folders only */}
-        {isFolder && (
-          <>
-            <div style={{ height: 1, background: 'var(--dome-border)', margin: '4px 6px' }} />
-            <button
-              type="button"
-              className="flex items-center w-full text-left transition-colors"
-              style={{
-                gap: 8, padding: '6px 12px', fontSize: 12.5, border: 'none', cursor: 'pointer',
-                color: 'var(--dome-text)',
-                background: showColors ? 'var(--dome-bg-hover)' : 'transparent',
-              }}
-              onMouseEnter={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'var(--dome-bg-hover)'; }}
-              onMouseLeave={(e) => { if (!showColors) (e.currentTarget as HTMLButtonElement).style.background = 'transparent'; }}
-              onClick={() => setShowColors((s) => !s)}
-            >
-              <span
-                className="size-3.5 rounded-full shrink-0 flex items-center justify-center"
-                style={{
-                  background: currentColor?.startsWith('#') ? currentColor : 'var(--dome-accent)',
-                  boxShadow: `0 0 0 1.5px ${currentColor?.startsWith('#') ? currentColor + '44' : 'transparent'}`,
-                }}
-              />
-              <span className="flex-1" style={{ fontWeight: 500 }}>Color de carpeta</span>
-              <ChevronDown
-                className="size-3 shrink-0"
-                style={{ color: 'var(--dome-text-muted)', transform: showColors ? 'rotate(0deg)' : 'rotate(-90deg)', transition: 'transform 150ms' }}
-              />
-            </button>
-
-            {showColors && (
-              <div style={{ padding: '6px 12px 10px' }}>
-                {/* Color grid */}
-                <div className="flex flex-wrap gap-1.5 mb-1.5">
-                  {FOLDER_COLOR_OPTIONS.map((opt) => {
-                    const isActive = currentColor === opt.value;
-                    return (
-                      <button
-                        key={opt.value}
-                        type="button"
-                        onClick={() => { onColorChange(r, opt.value); onClose(); }}
-                        onMouseEnter={() => setHoveredColor(opt.value)}
-                        onMouseLeave={() => setHoveredColor(null)}
-                        className="relative flex items-center justify-center transition-transform"
-                        style={{
-                          width: 22, height: 22, borderRadius: '50%',
-                          background: opt.value, border: 'none', cursor: 'pointer',
-                          outline: isActive ? `2.5px solid ${opt.value}` : '2px solid transparent',
-                          outlineOffset: isActive ? 2 : 0,
-                          transform: hoveredColor === opt.value ? 'scale(1.18)' : 'scale(1)',
-                          transition: 'transform 120ms, outline 120ms',
-                          boxShadow: hoveredColor === opt.value ? `0 2px 8px ${opt.value}66` : 'none',
-                        }}
-                      >
-                        {isActive && <Check className="size-2.5 text-white" strokeWidth={3} />}
-                      </button>
-                    );
-                  })}
-                </div>
-
-                {/* Color label tooltip */}
-                <div style={{ height: 16, display: 'flex', alignItems: 'center' }}>
-                  {hoveredLabel ? (
-                    <span style={{
-                      fontSize: 12, fontWeight: 500, color: hoveredColor ?? 'var(--dome-text-muted)',
-                      transition: 'color 100ms',
-                    }}>
-                      {hoveredLabel}
-                    </span>
-                  ) : currentColor ? (
-                    <span style={{ fontSize: 12, color: 'var(--dome-text-muted)' }}>
-                      {FOLDER_COLOR_OPTIONS.find((o) => o.value === currentColor)?.label ?? 'Personalizado'}
-                    </span>
-                  ) : (
-                    <span style={{ fontSize: 12, color: 'var(--dome-text-muted)' }}>{t('ui.no_color')}</span>
-                  )}
-                </div>
-              </div>
-            )}
-          </>
-        )}
-
-        <div style={{ height: 1, background: 'var(--dome-border)', margin: '4px 6px' }} />
-
-        {/* Delete */}
-        <CtxItem
-          icon={<Trash2 className="size-3.5" />}
-          label="Eliminar"
-          onClick={() => { onDelete(r); onClose(); }}
-          danger
-        />
-      </div>
-    </div>
-  );
-}
-
-// Shared context menu item
-function CtxItem({ icon, label, onClick, danger = false }: {
-  icon: React.ReactNode; label: string; onClick: () => void; danger?: boolean;
-}) {
-  const [hovered, setHovered] = useState(false);
-  return (
-    <button
-      type="button"
-      className="flex items-center w-full text-left transition-colors"
-      style={{
-        gap: 8, padding: '6px 12px', fontSize: 12.5, border: 'none', cursor: 'pointer',
-        color: danger ? 'var(--dome-error)' : 'var(--dome-text)',
-        background: hovered ? (danger ? 'rgba(239,68,68,0.08)' : 'var(--dome-bg-hover)') : 'transparent',
-        fontWeight: 450,
-      }}
-      onMouseEnter={() => setHovered(true)}
-      onMouseLeave={() => setHovered(false)}
-      onClick={onClick}
-    >
-      <span style={{ opacity: 0.75, color: danger ? 'var(--dome-error)' : 'var(--dome-text-muted)', display: 'flex' }}>{icon}</span>
-      <span>{label}</span>
-    </button>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Move folder modal
-// ---------------------------------------------------------------------------
-function MoveFolderModal({ resource, allFolders, onConfirm, onClose }: {
-  resource: Resource; allFolders: Resource[];
-  onConfirm: (folderId: string | null) => void; onClose: () => void;
-}) {
-  const { t } = useTranslation();
-  const [selected, setSelected] = useState<string | null>(resource.folder_id ?? null);
-  const available = allFolders.filter((f) => f.id !== resource.id);
-
-  return (
-    <div className="fixed inset-0 z-[9998] flex items-center justify-center" role="presentation">
-      <button
-        type="button"
-        className="absolute inset-0 min-h-full w-full cursor-pointer border-0 p-0"
-        style={{ background: 'rgba(0,0,0,0.45)' }}
-        aria-label={t('ui.close')}
-        onClick={onClose}
-      />
-      <div
-        className="relative z-10 rounded-xl shadow-xl border flex flex-col"
-        style={{ width: 300, maxHeight: 400, background: 'var(--dome-surface)', borderColor: 'var(--dome-border)' }}
-        role="dialog"
-        aria-modal="true"
-      >
-        <div className="flex items-center justify-between px-4 py-3 border-b" style={{ borderColor: 'var(--dome-border)' }}>
-          <span className="font-medium text-sm" style={{ color: 'var(--dome-text)' }}>Mover "{resource.title}"</span>
-          <button type="button" onClick={onClose} className="rounded flex items-center justify-center hover:bg-[var(--dome-bg-hover)]" style={{ width: 24, height: 24, background: 'none', border: 'none', cursor: 'pointer', color: 'var(--dome-text-muted)' }}>
-            <X className="size-3.5" />
-          </button>
-        </div>
-        <div className="flex-1 overflow-y-auto py-1.5">
-          <button type="button" className="flex items-center gap-2 w-full text-left px-4 py-2 transition-colors"
-            style={{ background: selected === null ? 'var(--dome-bg-hover)' : 'transparent', border: 'none', cursor: 'pointer', fontSize: 12.5, color: 'var(--dome-text)' }}
-            onClick={() => setSelected(null)}>
-            <Hash className="size-3.5 shrink-0" style={{ color: 'var(--dome-text-muted)' }} />
-            <span className="flex-1">{t('ui.no_folder_root')}</span>
-            {selected === null && <Check className="size-3.5 shrink-0" style={{ color: 'var(--dome-accent)' }} />}
-          </button>
-          {available.map((f) => (
-            <button key={f.id} type="button"
-              className="flex items-center gap-2 w-full text-left px-4 py-2 transition-colors"
-              style={{ background: selected === f.id ? 'var(--dome-bg-hover)' : 'transparent', border: 'none', cursor: 'pointer', fontSize: 12.5, color: 'var(--dome-text)' }}
-              onClick={() => setSelected(f.id)}
-              onMouseEnter={(e) => { if (selected !== f.id) (e.currentTarget as HTMLButtonElement).style.background = 'var(--dome-bg-hover)'; }}
-              onMouseLeave={(e) => { if (selected !== f.id) (e.currentTarget as HTMLButtonElement).style.background = 'transparent'; }}>
-              <Folder className="size-3.5 shrink-0" style={{ color: getFolderColor(f) }} />
-              <span className="flex-1 truncate">{f.title}</span>
-              {selected === f.id && <Check className="size-3.5 shrink-0" style={{ color: 'var(--dome-accent)' }} />}
-            </button>
-          ))}
-        </div>
-        <div className="flex items-center justify-end gap-2 px-4 py-3 border-t" style={{ borderColor: 'var(--dome-border)' }}>
-          <button type="button" onClick={onClose} className="px-3 py-1.5 rounded-md text-xs" style={{ background: 'var(--dome-bg-hover)', border: 'none', cursor: 'pointer', color: 'var(--dome-text-muted)' }}>{t('common.cancel')}</button>
-          <button type="button" onClick={() => { onConfirm(selected); onClose(); }} className="px-3 py-1.5 rounded-md text-xs font-medium" style={{ background: 'var(--dome-accent)', border: 'none', cursor: 'pointer', color: 'var(--base-text)' }}>{t('common.move')}</button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Delete confirm modal
-// ---------------------------------------------------------------------------
-function DeleteConfirmModal({ resource, onConfirm, onClose }: {
-  resource: Resource; onConfirm: () => void; onClose: () => void;
-}) {
-  const { t } = useTranslation();
-  return (
-    <div className="fixed inset-0 z-[9998] flex items-center justify-center" role="presentation">
-      <button
-        type="button"
-        className="absolute inset-0 min-h-full w-full cursor-pointer border-0 p-0"
-        style={{ background: 'rgba(0,0,0,0.45)' }}
-        aria-label={t('ui.close')}
-        onClick={onClose}
-      />
-      <div
-        className="relative z-10 rounded-xl shadow-xl border p-5 flex flex-col gap-3"
-        style={{ width: 290, background: 'var(--dome-surface)', borderColor: 'var(--dome-border)' }}
-        role="dialog"
-        aria-modal="true"
-      >
-        <div>
-          <p className="font-medium text-sm mb-1" style={{ color: 'var(--dome-text)' }}>
-            {t('ui.delete_confirm', { type: resource.type === 'folder' ? 'folder' : 'resource' })}
-          </p>
-          <p className="text-xs" style={{ color: 'var(--dome-text-muted)' }}>
-            {resource.type === 'folder' ? t('ui.delete_content_warning') : t('ui.delete_warning')}
-          </p>
-        </div>
-        <div className="flex items-center justify-end gap-2">
-          <button type="button" onClick={onClose} className="px-3 py-1.5 rounded-md text-xs" style={{ background: 'var(--dome-bg-hover)', border: 'none', cursor: 'pointer', color: 'var(--dome-text-muted)' }}>{t('ui.cancel')}</button>
-          <button type="button" onClick={() => { onConfirm(); onClose(); }} className="px-3 py-1.5 rounded-md text-xs font-medium" style={{ background: 'var(--dome-error)', border: 'none', cursor: 'pointer', color: 'var(--base-text)' }}>{t('ui.delete')}</button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// New folder modal
-// ---------------------------------------------------------------------------
-function NewFolderModal({ parentId, onConfirm, onClose }: {
-  parentId: string | null; onConfirm: (name: string, parentId: string | null) => void; onClose: () => void;
-}) {
-  const { t } = useTranslation();
-  const [name, setName] = useState('');
-  const inputRef = useRef<HTMLInputElement>(null);
-  useEffect(() => { inputRef.current?.focus(); }, []);
-
-  const submit = () => { const t = name.trim(); if (t) { onConfirm(t, parentId); onClose(); } };
-
-  return (
-    <div className="fixed inset-0 z-[9998] flex items-center justify-center" role="presentation">
-      <button
-        type="button"
-        className="absolute inset-0 min-h-full w-full cursor-pointer border-0 p-0"
-        style={{ background: 'rgba(0,0,0,0.45)' }}
-        aria-label={t('ui.close')}
-        onClick={onClose}
-      />
-      <div
-        className="relative z-10 rounded-xl shadow-xl border p-5 flex flex-col gap-3"
-        style={{ width: 280, background: 'var(--dome-surface)', borderColor: 'var(--dome-border)' }}
-        role="dialog"
-        aria-modal="true"
-      >
-        <p className="font-medium text-sm" style={{ color: 'var(--dome-text)' }}>{t('ui.new_folder')}</p>
-        <input ref={inputRef} type="text" value={name} onChange={(e) => setName(e.target.value)}
-          onKeyDown={(e) => { if (e.key === 'Enter') submit(); if (e.key === 'Escape') onClose(); }}
-          placeholder={t('ui.folder_name')} className="rounded-md px-3 py-2 text-sm outline-none"
-          style={{ background: 'var(--dome-bg-hover)', border: '1px solid var(--dome-border)', color: 'var(--dome-text)' }} />
-        <div className="flex items-center justify-end gap-2">
-          <button type="button" onClick={onClose} className="px-3 py-1.5 rounded-md text-xs" style={{ background: 'var(--dome-bg-hover)', border: 'none', cursor: 'pointer', color: 'var(--dome-text-muted)' }}>{t('ui.cancel')}</button>
-          <button type="button" onClick={submit} disabled={!name.trim()} className="px-3 py-1.5 rounded-md text-xs font-medium" style={{ background: 'var(--dome-accent)', border: 'none', cursor: 'pointer', color: 'var(--base-text)', opacity: name.trim() ? 1 : 0.5 }}>{t('ui.create')}</button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// TreeNode — individual row with hover menu + inline rename + drag-and-drop
-// ---------------------------------------------------------------------------
-interface TreeNodeProps {
-  node: TreeNodeData;
-  depth: number;
-  expandedIds: Set<string>;
-  onToggle: (id: string) => void;
-  onSelect: (node: TreeNodeData) => void;
-  onOpenFolder: (folderId: string, title: string) => void;
-  renameId: string | null;
-  dragOverId: string | null;
-  onContextMenu: (e: React.MouseEvent, r: Resource) => void;
-  onRenameCommit: (id: string, newTitle: string) => void;
-  onRenameCancel: () => void;
-  onDragStart: (node: TreeNodeData) => void;
-  onDragOver: (e: React.DragEvent, node: TreeNodeData) => void;
-  onDragLeave: () => void;
-  onDrop: (e: React.DragEvent, targetNode: TreeNodeData) => void;
-  onDragEnd: () => void;
-  /** Selection mode props */
-  selectedIds?: Set<string>;
-  onToggleSelect?: (id: string) => void;
-}
-
-function TreeNode({
-  node, depth, expandedIds, onToggle, onSelect, onOpenFolder, renameId, dragOverId,
-  onContextMenu, onRenameCommit, onRenameCancel,
-  onDragStart, onDragOver, onDragLeave, onDrop, onDragEnd,
-  selectedIds, onToggleSelect,
-}: TreeNodeProps) {
-  const isFolder = node.type === 'folder';
-  const isExpanded = expandedIds.has(node.id);
-  const hasChildren = isFolder && node.children && node.children.length > 0;
-  const isRenaming = renameId === node.id;
-  const isDragOver = dragOverId === node.id;
-  const [hovered, setHovered] = useState(false);
-  const [renameValue, setRenameValue] = useState(node.name);
-  const renameRef = useRef<HTMLInputElement>(null);
-
-  useEffect(() => {
-    if (isRenaming) { setRenameValue(node.name); const timer = setTimeout(() => renameRef.current?.select(), 10); return () => clearTimeout(timer); }
-  }, [isRenaming, node.name]);
-
-  const handleClick = () => {
-    if (isRenaming) return;
-    if (isFolder) {
-      onToggle(node.id);
-      onOpenFolder(node.id, node.name);
-    } else {
-      onSelect(node);
-    }
-  };
-
-  const handleRenameKeyDown = (e: React.KeyboardEvent) => {
-    if (e.key === 'Enter') { e.preventDefault(); onRenameCommit(node.id, renameValue); }
-    if (e.key === 'Escape') onRenameCancel();
-  };
-
-  const folderColor = isFolder && node.resource ? getFolderColor(node.resource) : 'var(--dome-accent)';
-
-  const isSelected = selectedIds?.has(node.id) ?? false;
-  const inSelectionMode = Boolean(onToggleSelect);
-
-  let rowBg = 'transparent';
-  if (isSelected) rowBg = 'color-mix(in srgb, var(--dome-accent) 10%, transparent)';
-  else if (isDragOver && isFolder) rowBg = `${folderColor}22`;
-  else if (hovered) rowBg = 'var(--dome-bg-hover)';
-
-  return (
-    <div>
-      <div
-        className="flex items-center w-full relative rounded transition-colors"
-        style={{
-          paddingLeft: 8 + depth * 14,
-          paddingRight: 4,
-          height: 28,
-          background: rowBg,
-          minWidth: 0,
-          outline: isSelected ? '1px solid color-mix(in srgb, var(--dome-accent) 40%, transparent)' : isDragOver && isFolder ? `1.5px dashed ${folderColor}` : 'none',
-          outlineOffset: -1,
-        }}
-        draggable={!isRenaming && !inSelectionMode}
-        onDragStart={() => !inSelectionMode && onDragStart(node)}
-        onDragOver={(e) => !inSelectionMode && onDragOver(e, node)}
-        onDragLeave={onDragLeave}
-        onDrop={(e) => !inSelectionMode && onDrop(e, node)}
-        onDragEnd={onDragEnd}
-        onMouseEnter={() => setHovered(true)}
-        onMouseLeave={() => setHovered(false)}
-        onContextMenu={(e) => { if (node.resource && !inSelectionMode) { e.preventDefault(); onContextMenu(e, node.resource); } }}
-      >
-        {/* Selection checkbox — only shown in selection mode */}
-        {inSelectionMode && (
-          <button
-            type="button"
-            role="checkbox"
-            onClick={(e) => { e.stopPropagation(); onToggleSelect!(node.id); }}
-            className="shrink-0 flex items-center justify-center rounded mr-1 transition-colors"
-            style={{
-              width: 14, height: 14,
-              border: `1.5px solid ${isSelected ? 'var(--dome-accent)' : 'var(--dome-border)'}`,
-              background: isSelected ? 'var(--dome-accent)' : 'var(--dome-bg)',
-              flexShrink: 0,
-            }}
-            aria-checked={isSelected}
-          >
-            {isSelected && (
-              <svg width="8" height="8" viewBox="0 0 8 8" fill="none" aria-hidden>
-                <path d="M1.5 4L3.5 6L6.5 2" stroke="white" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" />
-              </svg>
-            )}
-          </button>
-        )}
-
-        <button
-          type="button"
-          onClick={inSelectionMode ? () => onToggleSelect!(node.id) : handleClick}
-          className="flex items-center flex-1 text-left min-w-0"
-          style={{ gap: 5, background: 'none', border: 'none', cursor: 'pointer', color: hovered ? 'var(--dome-text)' : 'var(--dome-text-secondary)', padding: 0, minWidth: 0 }}
-        >
-          <span className="shrink-0 flex items-center justify-center" style={{ width: 14, height: 14 }}>
-            {isFolder
-              ? <ChevronDown className={`size-3 transition-transform ${isExpanded ? '' : '-rotate-90'}`} strokeWidth={2.5} />
-              : null}
-          </span>
-
-          {/* Folder icon with color swatch */}
-          {isFolder ? (
-            <span className="shrink-0 relative flex items-center justify-center">
-              <Folder className="size-3.5" style={{ color: folderColor }} strokeWidth={1.75} fill={`${folderColor}33`} />
-            </span>
-          ) : (
-            <span className="shrink-0" style={{ color: 'var(--dome-text-muted)' }}>
-              <DomeResourceIcon type={node.type} name={node.name} size={14} className="size-3.5" strokeWidth={1.75} />
-            </span>
-          )}
-
-          {isRenaming ? (
-            <input
-              ref={renameRef}
-              type="text"
-              value={renameValue}
-              onChange={(e) => setRenameValue(e.target.value)}
-              onKeyDown={handleRenameKeyDown}
-              onBlur={() => onRenameCommit(node.id, renameValue)}
-              onClick={(e) => e.stopPropagation()}
-              className="flex-1 outline-none rounded px-1"
-              style={{ fontSize: 12, background: 'var(--dome-surface)', border: '1px solid var(--dome-accent)', color: 'var(--dome-text)', minWidth: 0 }}
-            />
-          ) : (
-            <span className="truncate flex-1" style={{ fontSize: 12, fontWeight: isFolder ? 500 : 400, color: isFolder ? folderColor : undefined }}>{node.name}</span>
-          )}
-        </button>
-
-        {hovered && !isRenaming && node.resource && (
-          <button
-            type="button"
-            onClick={(e) => { e.stopPropagation(); if (node.resource) onContextMenu(e, node.resource); }}
-            className="shrink-0 flex items-center justify-center rounded-md transition-all"
-            style={{ width: 20, height: 20, background: 'transparent', border: 'none', cursor: 'pointer', color: 'var(--dome-text-muted)', flexShrink: 0 }}
-            onMouseEnter={(e) => {
-              (e.currentTarget as HTMLButtonElement).style.background = 'var(--dome-bg)';
-              (e.currentTarget as HTMLButtonElement).style.color = 'var(--dome-text)';
-            }}
-            onMouseLeave={(e) => {
-              (e.currentTarget as HTMLButtonElement).style.background = 'transparent';
-              (e.currentTarget as HTMLButtonElement).style.color = 'var(--dome-text-muted)';
-            }}
-          >
-            <MoreHorizontal className="size-3.5" />
-          </button>
-        )}
-      </div>
-
-      {isExpanded && hasChildren && (
-        <div style={{ borderLeft: `1.5px solid ${folderColor}44`, marginLeft: 8 + depth * 14 + 7 }}>
-          {node.children!.map((child) => (
-            <TreeNode
-              key={child.id}
-              node={child}
-              depth={depth + 1}
-              expandedIds={expandedIds}
-              onToggle={onToggle}
-              onSelect={onSelect}
-              onOpenFolder={onOpenFolder}
-              renameId={renameId}
-              dragOverId={dragOverId}
-              onContextMenu={onContextMenu}
-              onRenameCommit={onRenameCommit}
-              onRenameCancel={onRenameCancel}
-              onDragStart={onDragStart}
-              onDragOver={onDragOver}
-              onDragLeave={onDragLeave}
-              onDrop={onDrop}
-              onDragEnd={onDragEnd}
-              selectedIds={selectedIds}
-              onToggleSelect={onToggleSelect}
-            />
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Helper: check if `target` is inside `ancestor` in the tree
-// ---------------------------------------------------------------------------
-function isDescendant(target: TreeNodeData, ancestor: TreeNodeData): boolean {
-  if (!ancestor.children) return false;
-  for (const child of ancestor.children) {
-    if (child.id === target.id || isDescendant(target, child)) return true;
-  }
-  return false;
-}
-
-// ---------------------------------------------------------------------------
-// FileTree — tree container with all state and operations
-// ---------------------------------------------------------------------------
-interface FileTreeProps {
-  resources: Resource[];
-  onRefresh: () => void;
-}
-
-function FileTree({ resources, onRefresh }: FileTreeProps) {
-  const { t } = useTranslation();
-  const [expandedIds, setExpandedIds] = useState<Set<string>>(new Set());
-  const [searchQuery, setSearchQuery] = useState('');
-  const [ctxMenu, setCtxMenu] = useState<CtxState>({ visible: false, x: 0, y: 0, resource: null });
-  const [renameId, setRenameId] = useState<string | null>(null);
-  const [moveResource, setMoveResource] = useState<Resource | null>(null);
-  const [deleteResource, setDeleteResource] = useState<Resource | null>(null);
-  const [newFolderParentId, setNewFolderParentId] = useState<string | null | undefined>(undefined);
-  const [dragNode, setDragNode] = useState<TreeNodeData | null>(null);
-  const [dragOverId, setDragOverId] = useState<string | null>(null);
-  const dragEnterCountRef = useRef<Record<string, number>>({});
-
-  // ── Multi-selection ────────────────────────────────────────────────────────
-  const [selectionMode, setSelectionMode] = useState(false);
-  const [selectedIds, setSelectedIds] = useState<Set<string>>(new Set());
-  const [bulkDeleteConfirm, setBulkDeleteConfirm] = useState(false);
-  const [bulkDeleting, setBulkDeleting] = useState(false);
-
-  const handleToggleSelect = useCallback((id: string) => {
-    setSelectedIds((prev) => {
-      const next = new Set(prev);
-      if (next.has(id)) next.delete(id); else next.add(id);
-      return next;
-    });
-  }, []);
-
-  const exitSelectionMode = useCallback(() => {
-    setSelectionMode(false);
-    setSelectedIds(new Set());
-    setBulkDeleteConfirm(false);
-  }, []);
-
-  const handleBulkDelete = useCallback(async () => {
-    if (selectedIds.size === 0) return;
-    setBulkDeleting(true);
-    try {
-      for (const id of selectedIds) {
-        await window.electron?.resource?.delete(id);
-      }
-      exitSelectionMode();
-      onRefresh();
-    } finally {
-      setBulkDeleting(false);
-    }
-  }, [selectedIds, exitSelectionMode, onRefresh]);
-
-  const { openResourceTab, openFolderTab } = useTabStore.getState();
-  const folders = resources.filter((r) => r.type === 'folder');
-
-  /**
-   * Whether there's an active tab capable of hosting a split. We avoid
-   * splitting the home tab because it has no primary resource.
-   */
-  const activeTabId = useTabStore((s) => s.activeTabId);
-  const tabs = useTabStore((s) => s.tabs);
-  const openResourceInSplit = useTabStore((s) => s.openResourceInSplit);
-  const canOpenInSplit = activeTabId !== null && activeTabId !== 'home' &&
-    Boolean(tabs.find((tb) => tb.id === activeTabId)?.resourceId);
-
-  const handleOpenInSplit = useCallback((r: Resource) => {
-    openResourceInSplit(r.id, r.type, r.title || '');
-  }, [openResourceInSplit]);
-
-  const handleOpenInWindow = useCallback(async (r: Resource) => {
-    if (!window.electron?.invoke) return;
-    if (r.type !== 'note') return;
-    try {
-      await window.electron.invoke('window:create', {
-        id: `note-focus:${r.id}`,
-        route: `/focus/note/${encodeURIComponent(r.id)}`,
-        options: {
-          width: 960,
-          height: 760,
-          minWidth: 560,
-          minHeight: 480,
-          title: `${r.title || 'Nota'} — Dome`,
-          transparent: false,
-        },
-      });
-    } catch (err) {
-      console.error('[UnifiedSidebar] Failed to open popout:', err);
-    }
-  }, []);
-
-  const handleToggle = useCallback((id: string) => {
-    setExpandedIds((prev) => { const next = new Set(prev); if (next.has(id)) next.delete(id); else next.add(id); return next; });
-  }, []);
-
-  const handleSelect = useCallback((node: TreeNodeData) => {
-    if (node.resource) openResourceTab(node.id, node.type, node.name);
-  }, [openResourceTab]);
-
-  const handleOpenFolder = useCallback((folderId: string, title: string) => {
-    openFolderTab(folderId, title);
-  }, [openFolderTab]);
-
-  const handleContextMenu = useCallback((e: React.MouseEvent, r: Resource) => {
-    e.preventDefault();
-    setCtxMenu({ visible: true, x: e.clientX, y: e.clientY, resource: r });
-  }, []);
-
-  const handleRenameCommit = useCallback(async (id: string, newTitle: string) => {
-    const trimmed = newTitle.trim();
-    if (!trimmed) { setRenameId(null); return; }
-    const r = resources.find((res) => res.id === id);
-    if (r && trimmed !== r.title) {
-      await window.electron?.db?.resources?.update({ id, title: trimmed, updated_at: Date.now() });
-      onRefresh();
-    }
-    setRenameId(null);
-  }, [resources, onRefresh]);
-
-  const handleMoveConfirm = useCallback(async (folderId: string | null) => {
-    if (!moveResource) return;
-    await window.electron?.db?.resources?.moveToFolder(moveResource.id, folderId);
-    setMoveResource(null);
-    onRefresh();
-  }, [moveResource, onRefresh]);
-
-  const handleColorChange = useCallback(async (r: Resource, color: string) => {
-    const meta = { ...parseMeta(r), color };
-    await window.electron?.db?.resources?.update({ id: r.id, metadata: JSON.stringify(meta), updated_at: Date.now() });
-    onRefresh();
-  }, [onRefresh]);
-
-  const handleDeleteConfirm = useCallback(async () => {
-    if (!deleteResource) return;
-    await window.electron?.resource?.delete(deleteResource.id);
-    setDeleteResource(null);
-    onRefresh();
-  }, [deleteResource, onRefresh]);
-
-  const handleNewFolderConfirm = useCallback(async (name: string, parentId: string | null) => {
-    const now = Date.now();
-    const result = await window.electron?.db?.resources?.create({
-      id: `res_${now}_${Math.random().toString(36).substr(2, 9)}`,
-      type: 'folder' as Resource['type'],
-      title: name,
-      folder_id: parentId,
-      project_id: (resources[0]?.project_id) || 'default',
-      metadata: { color: pickFolderColor() },
-      created_at: now,
-      updated_at: now,
-    });
-    if (result?.success) {
-      if (parentId) setExpandedIds((prev) => new Set(prev).add(parentId));
-      onRefresh();
-    }
-  }, [resources, onRefresh]);
-
-  // Drag-and-drop handlers
-  const handleDragStart = useCallback((node: TreeNodeData) => {
-    setDragNode(node);
-    dragEnterCountRef.current = {};
-  }, []);
-
-  const handleDragOver = useCallback((e: React.DragEvent, target: TreeNodeData) => {
-    if (!dragNode) return;
-    if (dragNode.id === target.id) return;
-    // Only allow drop on folders; also allow drop on root (handled separately)
-    if (target.type !== 'folder') return;
-    e.preventDefault();
-    e.dataTransfer.dropEffect = 'move';
-    if (dragOverId !== target.id) setDragOverId(target.id);
-    // Auto-expand folder after hovering
-    setExpandedIds((prev) => new Set(prev).add(target.id));
-  }, [dragNode, dragOverId]);
-
-  const handleDragLeave = useCallback(() => {
-    setDragOverId(null);
-  }, []);
-
-  const handleDrop = useCallback(async (e: React.DragEvent, target: TreeNodeData) => {
-    e.preventDefault();
-    if (!dragNode || !dragNode.resource) { setDragOverId(null); setDragNode(null); return; }
-    if (dragNode.id === target.id) { setDragOverId(null); setDragNode(null); return; }
-    if (target.type !== 'folder') { setDragOverId(null); setDragNode(null); return; }
-    // Don't move a folder into itself or its descendant
-    if (dragNode.type === 'folder' && isDescendant(target, dragNode)) { setDragOverId(null); setDragNode(null); return; }
-    await window.electron?.db?.resources?.moveToFolder(dragNode.id, target.id);
-    setDragNode(null);
-    setDragOverId(null);
-    onRefresh();
-  }, [dragNode, onRefresh]);
-
-  const handleDragEnd = useCallback(() => {
-    setDragNode(null);
-    setDragOverId(null);
-  }, []);
-
-  const tree = buildTree(resources);
-  const q = searchQuery.trim().toLowerCase();
-
-  const filterTree = (nodes: TreeNodeData[]): TreeNodeData[] => {
-    if (!q) return nodes;
-    return nodes.reduce<TreeNodeData[]>((acc, node) => {
-      const matches = node.name.toLowerCase().includes(q);
-      const filteredChildren = node.children ? filterTree(node.children) : undefined;
-      if (matches || (filteredChildren && filteredChildren.length > 0)) acc.push({ ...node, children: filteredChildren });
-      return acc;
-    }, []);
-  };
-
-  const filteredTree = filterTree(tree);
-
-  return (
-    <div className="flex flex-col h-full">
-      {/* Search + selection toggle */}
-      <div className="px-3 pt-2 pb-1.5 flex items-center gap-1.5">
-        <div className="flex-1 flex items-center gap-1.5 rounded px-2" style={{ height: 26, background: 'var(--dome-bg-hover)', border: '1px solid var(--dome-border)' }}>
-          <Search className="size-3 shrink-0" style={{ color: 'var(--dome-text-muted)' }} strokeWidth={2} />
-          <input
-            type="text" value={searchQuery} onChange={(e) => setSearchQuery(e.target.value)}
-            placeholder={t('workspace.search_resources')} className="flex-1 bg-transparent outline-none border-none"
-            style={{ fontSize: 12, color: 'var(--dome-text)', caretColor: 'var(--dome-accent)' }}
-          />
-        </div>
-        {!selectionMode ? (
-          <button
-            type="button"
-            title={t('common.select')}
-            onClick={() => setSelectionMode(true)}
-            className="shrink-0 flex items-center justify-center rounded"
-            style={{ width: 24, height: 26, background: 'var(--dome-bg-hover)', border: '1px solid var(--dome-border)', cursor: 'pointer', color: 'var(--dome-text-muted)' }}
-          >
-            <Check className="size-3" />
-          </button>
-        ) : (
-          <button
-            type="button"
-            title={t('common.cancel')}
-            onClick={exitSelectionMode}
-            className="shrink-0 flex items-center justify-center rounded"
-            style={{ width: 24, height: 26, background: 'color-mix(in srgb, var(--dome-accent) 12%, transparent)', border: '1px solid var(--dome-accent)', cursor: 'pointer', color: 'var(--dome-accent)' }}
-          >
-            <X className="size-3" />
-          </button>
-        )}
-      </div>
-
-      {/* Selection action bar */}
-      {selectionMode && (
-        <div className="px-3 pb-1.5 flex items-center gap-1.5">
-          <span className="flex-1 text-[12px]" style={{ color: 'var(--dome-text-muted)' }}>
-            {selectedIds.size > 0 ? `${selectedIds.size} sel.` : t('common.select')}
-          </span>
-          {selectedIds.size > 0 && (
-            <button
-              type="button"
-              onClick={() => setBulkDeleteConfirm(true)}
-              disabled={bulkDeleting}
-              className="flex items-center gap-1 rounded px-2 py-0.5 text-[12px] font-medium"
-              style={{
-                background: 'color-mix(in srgb, var(--dome-error) 10%, transparent)',
-                color: 'var(--dome-error)',
-                border: 'none',
-                cursor: 'pointer',
-              }}
-            >
-              <Trash2 className="size-3" />
-              {t('common.delete')}
-            </button>
-          )}
-        </div>
-      )}
-
-      {/* Tree */}
-      <div className="flex-1 overflow-y-auto px-2 pb-2">
-        {filteredTree.length === 0 ? (
-          <p className="text-center py-4" style={{ fontSize: 12, color: 'var(--dome-text-muted)' }}>
-            {t('ui.no_results')}
-          </p>
-        ) : (
-          filteredTree.map((node) => (
-            <TreeNode
-              key={node.id}
-              node={node}
-              depth={0}
-              expandedIds={expandedIds}
-              onToggle={handleToggle}
-              onSelect={handleSelect}
-              onOpenFolder={handleOpenFolder}
-              renameId={renameId}
-              dragOverId={dragOverId}
-              onContextMenu={handleContextMenu}
-              onRenameCommit={handleRenameCommit}
-              onRenameCancel={() => setRenameId(null)}
-              onDragStart={handleDragStart}
-              onDragOver={handleDragOver}
-              onDragLeave={handleDragLeave}
-              onDrop={handleDrop}
-              onDragEnd={handleDragEnd}
-              selectedIds={selectionMode ? selectedIds : undefined}
-              onToggleSelect={selectionMode ? handleToggleSelect : undefined}
-            />
-          ))
-        )}
-      </div>
-
-      {/* Context menu */}
-      <ContextMenu
-        state={ctxMenu}
-        onClose={() => setCtxMenu((s) => ({ ...s, visible: false }))}
-        onRename={(r) => setRenameId(r.id)}
-        onMove={(r) => setMoveResource(r)}
-        onColorChange={handleColorChange}
-        onDelete={(r) => setDeleteResource(r)}
-        onNewFolder={(parentId) => setNewFolderParentId(parentId)}
-        onOpenInSplit={handleOpenInSplit}
-        onOpenInWindow={(r) => { void handleOpenInWindow(r); }}
-        canOpenInSplit={canOpenInSplit}
-      />
-
-      {moveResource && (
-        <MoveFolderModal resource={moveResource} allFolders={folders}
-          onConfirm={handleMoveConfirm} onClose={() => setMoveResource(null)} />
-      )}
-      {deleteResource && (
-        <DeleteConfirmModal resource={deleteResource}
-          onConfirm={handleDeleteConfirm} onClose={() => setDeleteResource(null)} />
-      )}
-      {newFolderParentId !== undefined && (
-        <NewFolderModal parentId={newFolderParentId}
-          onConfirm={handleNewFolderConfirm} onClose={() => setNewFolderParentId(undefined)} />
-      )}
-
-      {/* Bulk delete confirm modal */}
-      {bulkDeleteConfirm && (
-        <div className="fixed inset-0 z-[9999] flex items-center justify-center" style={{ background: 'rgba(0,0,0,0.45)' }}>
-          <div className="rounded-xl shadow-xl border p-4 flex flex-col gap-3" style={{ width: 270, background: 'var(--dome-surface)', borderColor: 'var(--dome-border)' }}>
-            <p className="font-medium text-sm" style={{ color: 'var(--dome-text)' }}>
-              {t('ui.delete_confirm', { type: 'items' })}
-            </p>
-            <p className="text-xs" style={{ color: 'var(--dome-text-muted)' }}>
-              {selectedIds.size} {t('common.select')}
-            </p>
-            <div className="flex items-center justify-end gap-2">
-              <button type="button" onClick={() => setBulkDeleteConfirm(false)}
-                className="px-3 py-1.5 rounded-md text-xs"
-                style={{ background: 'var(--dome-bg-hover)', border: 'none', cursor: 'pointer', color: 'var(--dome-text-muted)' }}>
-                {t('ui.cancel')}
-              </button>
-              <button
-                type="button"
-                disabled={bulkDeleting}
-                onClick={() => void handleBulkDelete()}
-                className="px-3 py-1.5 rounded-md text-xs font-medium"
-                style={{ background: 'var(--dome-error)', border: 'none', cursor: 'pointer', color: 'var(--base-text)', opacity: bulkDeleting ? 0.6 : 1 }}>
-                {bulkDeleting ? '...' : t('ui.delete')}
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// URL input modal
-// ---------------------------------------------------------------------------
-function UrlInputModal({ onConfirm, onClose }: { onConfirm: (url: string) => void; onClose: () => void }) {
-  const { t } = useTranslation();
-  const [url, setUrl] = useState('');
-  const inputRef = useRef<HTMLInputElement>(null);
-  useEffect(() => { inputRef.current?.focus(); }, []);
-
-  const submit = () => {
-    const t = url.trim();
-    if (t) { onConfirm(t); onClose(); }
-  };
-
-  return (
-    <div className="fixed inset-0 z-[9998] flex items-center justify-center" role="presentation">
-      <button
-        type="button"
-        className="absolute inset-0 min-h-full w-full cursor-pointer border-0 p-0"
-        style={{ background: 'rgba(0,0,0,0.45)' }}
-        aria-label={t('ui.close')}
-        onClick={onClose}
-      />
-      <div
-        className="relative z-10 rounded-xl shadow-xl border p-5 flex flex-col gap-3"
-        style={{ width: 320, background: 'var(--dome-surface)', borderColor: 'var(--dome-border)' }}
-        role="dialog"
-        aria-modal="true"
-      >
-        <p className="font-medium text-sm" style={{ color: 'var(--dome-text)' }}>{t('ui.add_url')}</p>
-        <input
-          ref={inputRef}
-          type="url"
-          value={url}
-          onChange={(e) => setUrl(e.target.value)}
-          onKeyDown={(e) => { if (e.key === 'Enter') submit(); if (e.key === 'Escape') onClose(); }}
-          placeholder="https://..."
-          className="rounded-md px-3 py-2 text-sm outline-none"
-          style={{ background: 'var(--dome-bg-hover)', border: '1px solid var(--dome-border)', color: 'var(--dome-text)' }}
-        />
-        <div className="flex items-center justify-end gap-2">
-          <button type="button" onClick={onClose} className="px-3 py-1.5 rounded-md text-xs" style={{ background: 'var(--dome-bg-hover)', border: 'none', cursor: 'pointer', color: 'var(--dome-text-muted)' }}>{t('ui.cancel')}</button>
-          <button type="button" onClick={submit} disabled={!url.trim()} className="px-3 py-1.5 rounded-md text-xs font-medium" style={{ background: 'var(--dome-accent)', border: 'none', cursor: 'pointer', color: 'var(--base-text)', opacity: url.trim() ? 1 : 0.5 }}>{t('ui.add')}</button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Add-resource dropdown
-// ---------------------------------------------------------------------------
-interface AddResourceMenuProps {
-  x: number;
-  y: number;
-  onClose: () => void;
-  onCreateNote: () => void;
-  onCreateNotebook: () => void;
-  onAddUrl: () => void;
-  onImportFile: () => void;
-  onImportFromCloud: () => void;
-  onCreateArtifact: () => void;
-}
-
-function AddResourceMenu({ x, y, onClose, onCreateNote, onCreateNotebook, onAddUrl, onImportFile, onImportFromCloud, onCreateArtifact }: AddResourceMenuProps) {
-  const { t } = useTranslation();
-  const menuRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const handle = (e: MouseEvent) => { if (menuRef.current && !menuRef.current.contains(e.target as Node)) onClose(); };
-    const handleKey = (e: KeyboardEvent) => { if (e.key === 'Escape') onClose(); };
-    document.addEventListener('mousedown', handle);
-    document.addEventListener('keydown', handleKey);
-    return () => { document.removeEventListener('mousedown', handle); document.removeEventListener('keydown', handleKey); };
-  }, [onClose]);
-
-  const itemStyle: React.CSSProperties = {
-    display: 'flex', alignItems: 'center', gap: 8, width: '100%',
-    textAlign: 'left', padding: '7px 12px', fontSize: 12.5,
-    color: 'var(--dome-text)', background: 'transparent', border: 'none', cursor: 'pointer',
-  };
-
-  const ITEMS = [
-    { icon: <FileText className="size-3.5" strokeWidth={1.75} />, label: t('toolbar.note'), action: onCreateNote },
-    { icon: <NotebookPen className="size-3.5" strokeWidth={1.75} />, label: 'Notebook', action: onCreateNotebook },
-    { icon: <Layers className="size-3.5" strokeWidth={1.75} />, label: t('artifacts.new_artifact'), action: onCreateArtifact },
-    { icon: <Link className="size-3.5" strokeWidth={1.75} />, label: 'URL / Artículo', action: onAddUrl },
-    { icon: <Upload className="size-3.5" strokeWidth={1.75} />, label: 'Importar fichero', action: onImportFile },
-    { icon: <Cloud className="size-3.5" strokeWidth={1.75} />, label: 'Importar desde Cloud', action: onImportFromCloud },
-  ];
-
-  return (
-    <div
-      ref={menuRef}
-      className="fixed z-[var(--z-popover)] rounded-lg shadow-xl border overflow-hidden"
-      style={{
-        left: Math.min(x, window.innerWidth - 200),
-        top: Math.min(y, window.innerHeight - 160),
-        minWidth: 170,
-        background: 'var(--dome-surface)',
-        borderColor: 'var(--dome-border)',
-        padding: '4px 0',
-      }}
-    >
-      {ITEMS.map((item) => (
-        <button
-          key={item.label}
-          style={itemStyle}
-          onMouseEnter={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'var(--dome-bg-hover)'; }}
-          onMouseLeave={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'transparent'; }}
-          onClick={() => { item.action(); onClose(); }}
-        >
-          <span style={{ color: 'var(--dome-text-muted)' }}>{item.icon}</span>
-          <span>{item.label}</span>
-        </button>
-      ))}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// timeAgo helper
-// ---------------------------------------------------------------------------
-function timeAgo(ts: number): string {
-  const diff = Date.now() - ts;
-  const seconds = Math.floor(diff / 1000);
-  if (seconds < 60) return `${seconds}s`;
-  const minutes = Math.floor(seconds / 60);
-  if (minutes < 60) return `${minutes}m`;
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) return `${hours}h`;
-  return `${Math.floor(hours / 24)}d`;
-}
-
-// Silence unused warning
-void timeAgo;
-
-// ---------------------------------------------------------------------------
-// UnifiedSidebar
-// ---------------------------------------------------------------------------
 interface UnifiedSidebarProps {
   collapsed: boolean;
   onCollapse: () => void;
